refactor: move BootInfo creation logic to FDT management module
