+++ kernel/src/arch/riscv64/boot/entry.rs
@@ -1,6 +1,6 @@
 use core::{arch::naked_asm, mem::transmute};
 
-use crate::{BootInfo, DeviceSource, arch::{Riscv64, riscv64::{TRAPFRAME, trap_init}}, device::fdt::{FdtManager, init_fdt, relocate_fdt}, environment::STACK_SIZE, mem::{__FDT_RESERVED_START, init_bss}, start_kernel, vm::vmem::MemoryArea};
+use crate::{arch::{Riscv64, riscv64::{TRAPFRAME, trap_init}}, device::fdt::{init_fdt, relocate_fdt, create_bootinfo_from_fdt}, environment::STACK_SIZE, mem::{__FDT_RESERVED_START, init_bss}, start_kernel};
 
 /// Entry point for the primary core
 #[unsafe(link_section = ".init")]
@@ -55,40 +55,7 @@ pub extern "C" fn _entry_ap() {
     }
 }
 
-fn create_bootinfo_from_fdt(cpu_id: usize, relocated_fdt_addr: usize) -> BootInfo {
-    let fdt_manager = FdtManager::get_manager();
-    
-    // Get DRAM area
-    let dram_area = fdt_manager.get_dram_memoryarea().expect("Memory area not found");
-    
-    // Calculate usable memory area (simplified for now)
-    let kernel_end = unsafe { &crate::mem::__KERNEL_SPACE_END as *const usize as usize };
-    let mut usable_memory = MemoryArea::new(kernel_end, dram_area.end);
-    
-    // Relocate initramfs
-    crate::early_println!("Relocating initramfs...");
-    
-    let relocated_initramfs = match crate::fs::vfs_v2::drivers::initramfs::relocate_initramfs(&mut usable_memory) {
-        Ok(area) => {
-            Some(area)
-        },
-        Err(e) => {
-            None
-        }
-    };
-    
-    // Get command line
-    let cmdline = fdt_manager.get_fdt()
-        .and_then(|fdt| fdt.chosen().bootargs());
-    
-    BootInfo::new(
-        cpu_id,
-        usable_memory,
-        relocated_initramfs,
-        cmdline,
-        DeviceSource::Fdt(relocated_fdt_addr),
-    )
-}
+
 
 #[unsafe(no_mangle)]
 pub extern "C" fn arch_start_kernel(hartid: usize, fdt_ptr: usize) {
